perf(things, items): improves performance on array accessors '[]'
a common use case is to use `items[dynamic_item_name]` or
`things[dynamic_thing_name]` to get a non-constant item or thing.

the old code would instantiate a new array or set of all items/things
in OpenHAB, and then immediately discard it when the [] method is used
to look up the item individually. this change delays instantiating
until an actual enumerable method is used

this also fixes

rule "some_rule" do
  changed GroupItem.members
end

querying and wrapping all the members of the group


Former-commit-id: 24bb04b34cc13ebbd4906fe547e31154d55e737c