feat(timers): support passing custom objects as timer duration
For example items such as `NumberItem` can be passed now. Basically
everything that responds to `#to_f` or `#to_i`. Output is evaluated as
number of seconds for timer.

Eg with a `Alarm_Delay` `NumberItem` this is now possible:

```rb
after Alarm_Delay { ... }

changed Alarm_Mode, for: Alarm_Delay
```

When rescheduling, the item's state will be revaluated.


Former-commit-id: 0fb0bcf0c7821f54685763b67b0f35c55de0f6f8