fix deadlock in specs
if a trigger executes outside the main thread, it would lock its
callback until the rule finished executing. if back in the main
thread, a spec is cleaning up, it would lock RuleEngine first,
and then try to clean up the rule. but the rule then finishes
executing, and cannot lock RuleEngine to mark it as done. deadlock!
fix it by _not_ running triggers synchronous unless the trigger
is actually coming from the main thread